requirements.txt
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
pytz==2018.9
